Now it can handle a model file.
Now you can load a 3D model file.
(This would just be, like, a .txt file. The file would be a list of points. All it needs to be is a list of numbers separated by commas... each group of 3 numbers gives you one 3D point. (And the program is set up, at the moment, to draw triangles. So, 3 points at a time (each with x, y, z) would probably be advisable.)